refactor: move StoreFb to ui/store.rs
