+++ crates/os/src/ui/mod.rs
@@ -1,4 +1,5 @@
 pub mod layout;
 pub mod logger;
+pub mod store;
 pub mod widget;
 pub mod writer;
